Extract item tree model to the model crate, keep item tree for lowering
