+++ crates/engine/analysis/Cargo.toml
@@ -19,7 +19,6 @@ rg_def_map.workspace = true
 rg_ir_model.workspace = true
 rg_ir_storage.workspace = true
 rg_ir_view.workspace = true
-rg_item_tree.workspace = true
 rg_parse.workspace = true
 rg_ty.workspace = true
