+++ crates/engine/ir-model/Cargo.toml
@@ -11,11 +11,13 @@ readme.workspace = true
 description.workspace = true
 
 [dependencies]
-derive_more = { workspace = true, features = ["from"] }
+derive_more = { workspace = true, features = ["display", "from"] }
 wincode.workspace = true
 rg_arena = { workspace = true, features = ["memsize"] }
+rg_cfg_eval.workspace = true
 rg_memsize.workspace = true
 rg_parse.workspace = true
 rg_workspace.workspace = true
-rg_item_tree.workspace = true
 rg_text.workspace = true
+rg_tt.workspace = true
+rg_wincode_utils.workspace = true
