+++ crates/engine/ir-view/Cargo.toml
@@ -17,7 +17,6 @@ rg_body_ir.workspace = true
 rg_def_map.workspace = true
 rg_ir_model.workspace = true
 rg_ir_storage.workspace = true
-rg_item_tree.workspace = true
 rg_package_store.workspace = true
 rg_parse.workspace = true
 rg_semantic_ir.workspace = true
