+++ crates/engine/ir-view/src/source/occurrence.rs
@@ -8,12 +8,12 @@ use rg_body_ir::{
     ValueReferenceSurface,
 };
 use rg_def_map::DefMapCursorCandidate;
+use rg_ir_model::items::FieldKey;
 use rg_ir_model::{
     BodyBindingRef, ModuleRef, TargetRef,
     identity::{DeclarationRef, ExprRef, FunctionBodyRef, LexicalScopeRef},
 };
 use rg_ir_storage::{Path, TypePathContext};
-use rg_item_tree::FieldKey;
 use rg_parse::{FileId, Span};
 use rg_semantic_ir::SemanticCursorCandidate;
