+++ crates/engine/item-tree/Cargo.toml
@@ -18,6 +18,7 @@ rayon.workspace = true
 wincode.workspace = true
 rg_arena = { workspace = true, features = ["memsize"] }
 rg_cfg_eval.workspace = true
+rg_ir_model.workspace = true
 rg_macro_expand.workspace = true
 rg_memsize.workspace = true
 rg_parse.workspace = true
