+++ crates/engine/item-tree/src/item/type_ref.rs
@@ -1,165 +1,28 @@
-use std::fmt;
-
+use rg_ir_model::items::{GenericArg, Mutability, TypeBound, TypePath, TypePathSegment, TypeRef};
+use rg_parse::{LineIndex, Span};
 use rg_syntax::{
     AstNode as _,
     ast::{self, HasGenericArgs},
 };
+use rg_text::NameInterner;
 
-use rg_parse::{LineIndex, Span};
-use rg_text::{Name, NameInterner};
+use super::{FromAst, normalized_syntax};
 
-use super::normalized_syntax;
+impl FromAst for TypeRef {
+    type AstNode = ast::Type;
+    type Context<'a> = (&'a LineIndex, &'a mut NameInterner);
 
-/// Syntax-level mutability marker used by lowered declarations and type refs.
-#[derive(
-    Debug,
-    Clone,
-    Copy,
-    PartialEq,
-    Eq,
-    wincode::SchemaRead,
-    wincode::SchemaWrite,
-    rg_memsize::MemorySize,
-)]
-#[memsize(leaf)]
-pub enum Mutability {
-    Shared,
-    Mutable,
-}
-
-impl Mutability {
-    pub fn from_mut_token(is_mut: bool) -> Self {
-        if is_mut { Self::Mutable } else { Self::Shared }
-    }
-}
-
-impl fmt::Display for Mutability {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Shared => write!(f, "shared"),
-            Self::Mutable => write!(f, "mut"),
-        }
-    }
-}
-
-/// Unresolved type syntax lowered into the item tree.
-///
-/// This intentionally stops before semantic resolution. `TypeRef` represents what the user wrote
-/// in an item declaration; resolving paths to definitions belongs to later IR layers.
-#[derive(
-    Debug, Clone, PartialEq, Eq, wincode::SchemaRead, wincode::SchemaWrite, rg_memsize::MemorySize,
-)]
-pub enum TypeRef {
-    Unknown(String),
-    Never,
-    Unit,
-    Infer,
-    Path(#[wincode(with = "rg_wincode_utils::WincodeDynamic<TypePath>")] TypePath),
-    Tuple(#[wincode(with = "rg_wincode_utils::WincodeDynamic<Vec<TypeRef>>")] Vec<TypeRef>),
-    Reference {
-        lifetime: Option<String>,
-        mutability: Mutability,
-        #[wincode(with = "rg_wincode_utils::WincodeDynamic<Box<TypeRef>>")]
-        inner: Box<TypeRef>,
-    },
-    RawPointer {
-        mutability: Mutability,
-        #[wincode(with = "rg_wincode_utils::WincodeDynamic<Box<TypeRef>>")]
-        inner: Box<TypeRef>,
-    },
-    Slice(#[wincode(with = "rg_wincode_utils::WincodeDynamic<Box<TypeRef>>")] Box<TypeRef>),
-    Array {
-        #[wincode(with = "rg_wincode_utils::WincodeDynamic<Box<TypeRef>>")]
-        inner: Box<TypeRef>,
-        len: Option<String>,
-    },
-    FnPointer {
-        #[wincode(with = "rg_wincode_utils::WincodeDynamic<Vec<TypeRef>>")]
-        params: Vec<TypeRef>,
-        #[wincode(with = "rg_wincode_utils::WincodeDynamic<Box<TypeRef>>")]
-        ret: Box<TypeRef>,
-    },
-    ImplTrait(#[wincode(with = "rg_wincode_utils::WincodeDynamic<Vec<TypeBound>>")] Vec<TypeBound>),
-    DynTrait(#[wincode(with = "rg_wincode_utils::WincodeDynamic<Vec<TypeBound>>")] Vec<TypeBound>),
-}
-
-impl TypeRef {
-    pub fn unknown_from_text(text: impl Into<String>) -> Self {
-        Self::Unknown(text.into())
-    }
-
-    /// Returns true when this type syntax is the special `Self` type.
-    pub fn is_self_type(&self) -> bool {
-        matches!(self, Self::Path(path) if path.is_self_type())
-    }
-
-    /// Returns the name of a plain single-segment type path.
-    pub fn type_param_name(&self) -> Option<Name> {
-        match self {
-            Self::Path(path) => path.single_name().cloned(),
-            _ => None,
-        }
-    }
-
-    /// Returns true when this type syntax contains explicit generic arguments anywhere inside it.
-    pub fn has_generic_args(&self) -> bool {
-        match self {
-            Self::Path(path) => path.segments.iter().any(|segment| !segment.args.is_empty()),
-            Self::Tuple(types) => types.iter().any(Self::has_generic_args),
-            Self::Reference { inner, .. }
-            | Self::RawPointer { inner, .. }
-            | Self::Slice(inner)
-            | Self::Array { inner, .. } => inner.has_generic_args(),
-            Self::FnPointer { params, ret } => {
-                params.iter().any(Self::has_generic_args) || ret.has_generic_args()
-            }
-            Self::ImplTrait(bounds) | Self::DynTrait(bounds) => {
-                bounds.iter().any(TypeBound::has_generic_args)
-            }
-            Self::Unknown(_) | Self::Never | Self::Unit | Self::Infer => false,
-        }
-    }
-
-    /// Returns true when this type syntax mentions one of the provided type parameter names.
-    pub fn mentions_type_param(&self, params: &[&str]) -> bool {
-        match self {
-            Self::Path(path) => path.segments.iter().any(|segment| {
-                params.contains(&segment.name.as_str())
-                    || segment
-                        .args
-                        .iter()
-                        .any(|arg| arg.mentions_type_param(params))
-            }),
-            Self::Tuple(types) => types.iter().any(|ty| ty.mentions_type_param(params)),
-            Self::Reference { inner, .. }
-            | Self::RawPointer { inner, .. }
-            | Self::Slice(inner)
-            | Self::Array { inner, .. } => inner.mentions_type_param(params),
-            Self::FnPointer {
-                params: fn_params,
-                ret,
-            } => {
-                fn_params.iter().any(|ty| ty.mentions_type_param(params))
-                    || ret.mentions_type_param(params)
-            }
-            Self::ImplTrait(bounds) | Self::DynTrait(bounds) => {
-                bounds.iter().any(|bound| bound.mentions_type_param(params))
-            }
-            Self::Unknown(_) | Self::Never | Self::Unit | Self::Infer => false,
-        }
-    }
-
-    pub fn from_ast(ty: ast::Type, line_index: &LineIndex, interner: &mut NameInterner) -> Self {
-        match ty {
+    fn from_ast(ty: &Self::AstNode, (line_index, interner): Self::Context<'_>) -> Self {
+        match ty.clone() {
             ast::Type::ArrayType(ty) => Self::Array {
                 inner: Box::new(
                     ty.ty()
-                        .map(|ty| Self::from_ast(ty, line_index, interner))
+                        .map(|ty| Self::from_ast(&ty, (line_index, &mut *interner)))
                         .unwrap_or_else(|| Self::unknown_from_text(normalized_syntax(&ty))),
                 ),
                 len: ty.const_arg().map(|arg| normalized_syntax(&arg)),
             },
-            ast::Type::DynTraitType(ty) => Self::DynTrait(TypeBound::list_from_ast(
+            ast::Type::DynTraitType(ty) => Self::DynTrait(type_bound_list_from_ast(
                 ty.type_bound_list(),
                 line_index,
                 interner,
@@ -172,22 +35,22 @@ impl TypeRef {
                     .map(|param| {
                         param
                             .ty()
-                            .map(|ty| Self::from_ast(ty, line_index, interner))
+                            .map(|ty| Self::from_ast(&ty, (line_index, &mut *interner)))
                             .unwrap_or_else(|| Self::Unknown(String::new()))
                     })
                     .collect(),
                 ret: Box::new(
                     ty.ret_type()
                         .and_then(|ret_ty| ret_ty.ty())
-                        .map(|ty| Self::from_ast(ty, line_index, interner))
+                        .map(|ty| Self::from_ast(&ty, (line_index, &mut *interner)))
                         .unwrap_or(Self::Unit),
                 ),
             },
             ast::Type::ForType(ty) => ty
                 .ty()
-                .map(|ty| Self::from_ast(ty, line_index, interner))
+                .map(|ty| Self::from_ast(&ty, (line_index, &mut *interner)))
                 .unwrap_or_else(|| Self::unknown_from_text(normalized_syntax(&ty))),
-            ast::Type::ImplTraitType(ty) => Self::ImplTrait(TypeBound::list_from_ast(
+            ast::Type::ImplTraitType(ty) => Self::ImplTrait(type_bound_list_from_ast(
                 ty.type_bound_list(),
                 line_index,
                 interner,
@@ -197,18 +60,18 @@ impl TypeRef {
             ast::Type::NeverType(_) => Self::Never,
             ast::Type::ParenType(ty) => ty
                 .ty()
-                .map(|ty| Self::from_ast(ty, line_index, interner))
+                .map(|ty| Self::from_ast(&ty, (line_index, &mut *interner)))
                 .unwrap_or_else(|| Self::unknown_from_text(normalized_syntax(&ty))),
             ast::Type::PathType(ty) => ty
                 .path()
-                .map(|path| TypePath::from_ast(path, line_index, interner))
+                .map(|path| TypePath::from_ast(&path, (line_index, &mut *interner)))
                 .map(Self::Path)
                 .unwrap_or_else(|| Self::unknown_from_text(normalized_syntax(&ty))),
             ast::Type::PtrType(ty) => Self::RawPointer {
                 mutability: Mutability::from_mut_token(ty.mut_token().is_some()),
                 inner: Box::new(
                     ty.ty()
-                        .map(|ty| Self::from_ast(ty, line_index, interner))
+                        .map(|ty| Self::from_ast(&ty, (line_index, &mut *interner)))
                         .unwrap_or_else(|| Self::unknown_from_text(normalized_syntax(&ty))),
                 ),
             },
@@ -217,19 +80,19 @@ impl TypeRef {
                 mutability: Mutability::from_mut_token(ty.mut_token().is_some()),
                 inner: Box::new(
                     ty.ty()
-                        .map(|ty| Self::from_ast(ty, line_index, interner))
+                        .map(|ty| Self::from_ast(&ty, (line_index, &mut *interner)))
                         .unwrap_or_else(|| Self::unknown_from_text(normalized_syntax(&ty))),
                 ),
             },
             ast::Type::SliceType(ty) => Self::Slice(Box::new(
                 ty.ty()
-                    .map(|ty| Self::from_ast(ty, line_index, interner))
+                    .map(|ty| Self::from_ast(&ty, (line_index, &mut *interner)))
                     .unwrap_or_else(|| Self::unknown_from_text(normalized_syntax(&ty))),
             )),
             ast::Type::TupleType(ty) => {
                 let fields = ty
                     .fields()
-                    .map(|ty| Self::from_ast(ty, line_index, interner))
+                    .map(|ty| Self::from_ast(&ty, (line_index, &mut *interner)))
                     .collect::<Vec<_>>();
                 if fields.is_empty() {
                     Self::Unit
@@ -239,144 +102,19 @@ impl TypeRef {
             }
         }
     }
-
-    pub fn shrink_to_fit(&mut self) {
-        match self {
-            Self::Unknown(text) => text.shrink_to_fit(),
-            Self::Path(path) => path.shrink_to_fit(),
-            Self::Tuple(types) => {
-                types.shrink_to_fit();
-                for ty in types {
-                    ty.shrink_to_fit();
-                }
-            }
-            Self::Reference {
-                lifetime, inner, ..
-            } => {
-                if let Some(lifetime) = lifetime {
-                    lifetime.shrink_to_fit();
-                }
-                inner.shrink_to_fit();
-            }
-            Self::RawPointer { inner, .. } | Self::Slice(inner) => inner.shrink_to_fit(),
-            Self::Array { inner, len } => {
-                inner.shrink_to_fit();
-                if let Some(len) = len {
-                    len.shrink_to_fit();
-                }
-            }
-            Self::FnPointer { params, ret } => {
-                params.shrink_to_fit();
-                for param in params {
-                    param.shrink_to_fit();
-                }
-                ret.shrink_to_fit();
-            }
-            Self::ImplTrait(bounds) | Self::DynTrait(bounds) => {
-                bounds.shrink_to_fit();
-                for bound in bounds {
-                    bound.shrink_to_fit();
-                }
-            }
-            Self::Never | Self::Unit | Self::Infer => {}
-        }
-    }
 }
 
-impl rg_memsize::Shrink for TypeRef {
-    fn shrink_to_fit(&mut self) {
-        TypeRef::shrink_to_fit(self);
-    }
-}
+impl FromAst for TypePath {
+    type AstNode = ast::Path;
+    type Context<'a> = (&'a LineIndex, &'a mut NameInterner);
 
-impl fmt::Display for TypeRef {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Unknown(text) if text.is_empty() => write!(f, "<unknown>"),
-            Self::Unknown(text) => write!(f, "<unsupported:{text}>"),
-            Self::Never => write!(f, "!"),
-            Self::Unit => write!(f, "()"),
-            Self::Infer => write!(f, "_"),
-            Self::Path(path) => write!(f, "{path}"),
-            Self::Tuple(types) => {
-                write!(f, "(")?;
-                for (idx, ty) in types.iter().enumerate() {
-                    if idx > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "{ty}")?;
-                }
-                if types.len() == 1 {
-                    write!(f, ",")?;
-                }
-                write!(f, ")")
-            }
-            Self::Reference {
-                lifetime,
-                mutability,
-                inner,
-            } => {
-                write!(f, "&")?;
-                if let Some(lifetime) = lifetime {
-                    write!(f, "{lifetime} ")?;
-                }
-                if matches!(mutability, Mutability::Mutable) {
-                    write!(f, "mut ")?;
-                }
-                write!(f, "{inner}")
-            }
-            Self::RawPointer { mutability, inner } => match mutability {
-                Mutability::Shared => write!(f, "*const {inner}"),
-                Mutability::Mutable => write!(f, "*mut {inner}"),
-            },
-            Self::Slice(inner) => write!(f, "[{inner}]"),
-            Self::Array { inner, len } => {
-                write!(f, "[{inner}; ")?;
-                match len {
-                    Some(len) => write!(f, "{len}")?,
-                    None => write!(f, "<unknown>")?,
-                }
-                write!(f, "]")
-            }
-            Self::FnPointer { params, ret } => {
-                write!(f, "fn(")?;
-                for (idx, param) in params.iter().enumerate() {
-                    if idx > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "{param}")?;
-                }
-                write!(f, ")")?;
-                if !matches!(ret.as_ref(), TypeRef::Unit) {
-                    write!(f, " -> {ret}")?;
-                }
-                Ok(())
-            }
-            Self::ImplTrait(bounds) => write_bounds(f, "impl ", bounds),
-            Self::DynTrait(bounds) => write_bounds(f, "dyn ", bounds),
-        }
-    }
-}
-
-#[derive(
-    Debug, Clone, PartialEq, Eq, wincode::SchemaRead, wincode::SchemaWrite, rg_memsize::MemorySize,
-)]
-pub struct TypePath {
-    /// Full source range of the path syntax, including separators around segments.
-    pub source_span: Span,
-    pub absolute: bool,
-    #[wincode(with = "rg_wincode_utils::WincodeDynamic<Vec<TypePathSegment>>")]
-    pub segments: Vec<TypePathSegment>,
-}
-
-impl TypePath {
-    pub fn from_ast(path: ast::Path, line_index: &LineIndex, interner: &mut NameInterner) -> Self {
+    fn from_ast(path: &Self::AstNode, (line_index, interner): Self::Context<'_>) -> Self {
         let source_span = Span::from_text_range(path.syntax().text_range());
         let absolute = path
             .first_segment()
             .is_some_and(|segment| segment.coloncolon_token().is_some());
         let mut segments = Vec::new();
-        Self::collect_segments(&path, line_index, interner, &mut segments);
+        collect_segments(&path, line_index, interner, &mut segments);
 
         Self {
             source_span,
@@ -384,156 +122,65 @@ impl TypePath {
             segments,
         }
     }
+}
 
-    fn collect_segments(
-        path: &ast::Path,
-        line_index: &LineIndex,
-        interner: &mut NameInterner,
-        segments: &mut Vec<TypePathSegment>,
-    ) {
-        if let Some(qualifier) = path.qualifier() {
-            Self::collect_segments(&qualifier, line_index, interner, segments);
-        }
-
-        if let Some(segment) = path.segment() {
-            segments.push(TypePathSegment::from_ast(&segment, line_index, interner));
-        }
+fn collect_segments(
+    path: &ast::Path,
+    line_index: &LineIndex,
+    interner: &mut NameInterner,
+    segments: &mut Vec<TypePathSegment>,
+) {
+    if let Some(qualifier) = path.qualifier() {
+        collect_segments(&qualifier, line_index, interner, segments);
     }
 
-    /// Returns the name of a single-segment relative path.
-    pub fn single_name(&self) -> Option<&Name> {
-        if self.absolute || self.segments.len() != 1 {
-            return None;
-        }
-
-        self.segments.first().map(|segment| &segment.name)
-    }
-
-    pub fn is_self_type(&self) -> bool {
-        self.single_name()
-            .is_some_and(|name| name.as_str() == "Self")
-    }
-
-    pub fn shrink_to_fit(&mut self) {
-        self.segments.shrink_to_fit();
-        for segment in &mut self.segments {
-            segment.shrink_to_fit();
-        }
+    if let Some(segment) = path.segment() {
+        segments.push(type_path_segment_from_ast(&segment, line_index, interner));
     }
 }
 
-impl fmt::Display for TypePath {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        if self.absolute {
-            write!(f, "::")?;
-        }
+fn type_path_segment_from_ast(
+    segment: &ast::PathSegment,
+    line_index: &LineIndex,
+    interner: &mut NameInterner,
+) -> TypePathSegment {
+    let name = segment
+        .name_ref()
+        .map(|name| interner.intern(name.syntax().text().to_string().trim()))
+        .unwrap_or_else(|| interner.intern(normalized_syntax(segment)));
+    let span = segment
+        .name_ref()
+        .map(|name| name.syntax().text_range())
+        .unwrap_or_else(|| segment.syntax().text_range());
+    let mut args = Vec::new();
 
-        for (idx, segment) in self.segments.iter().enumerate() {
-            if idx > 0 {
-                write!(f, "::")?;
-            }
-            write!(f, "{segment}")?;
-        }
+    if let Some(arg_list) = segment.generic_arg_list() {
+        args.extend(
+            arg_list
+                .generic_args()
+                .map(|arg| GenericArg::from_ast(&arg, (line_index, &mut *interner))),
+        );
+    }
 
-        Ok(())
+    if let Some(parenthesized_args) = segment.parenthesized_arg_list() {
+        args.push(GenericArg::Unsupported(normalized_syntax(
+            &parenthesized_args,
+        )));
+    }
+
+    TypePathSegment {
+        name,
+        args,
+        span: Span::from_text_range(span),
     }
 }
 
-#[derive(
-    Debug, Clone, PartialEq, Eq, wincode::SchemaRead, wincode::SchemaWrite, rg_memsize::MemorySize,
-)]
-pub struct TypePathSegment {
-    pub name: Name,
-    #[wincode(with = "rg_wincode_utils::WincodeDynamic<Vec<GenericArg>>")]
-    pub args: Vec<GenericArg>,
-    pub span: Span,
-}
+impl FromAst for GenericArg {
+    type AstNode = ast::GenericArg;
+    type Context<'a> = (&'a LineIndex, &'a mut NameInterner);
 
-impl TypePathSegment {
-    fn from_ast(
-        segment: &ast::PathSegment,
-        line_index: &LineIndex,
-        interner: &mut NameInterner,
-    ) -> Self {
-        let name = segment
-            .name_ref()
-            .map(|name| interner.intern(name.syntax().text().to_string().trim()))
-            .unwrap_or_else(|| interner.intern(normalized_syntax(segment)));
-        let span = segment
-            .name_ref()
-            .map(|name| name.syntax().text_range())
-            .unwrap_or_else(|| segment.syntax().text_range());
-        let mut args = Vec::new();
-
-        if let Some(arg_list) = segment.generic_arg_list() {
-            args.extend(
-                arg_list
-                    .generic_args()
-                    .map(|arg| GenericArg::from_ast(arg, line_index, interner)),
-            );
-        }
-
-        if let Some(parenthesized_args) = segment.parenthesized_arg_list() {
-            args.push(GenericArg::Unsupported(normalized_syntax(
-                &parenthesized_args,
-            )));
-        }
-
-        Self {
-            name,
-            args,
-            span: Span::from_text_range(span),
-        }
-    }
-
-    pub fn shrink_to_fit(&mut self) {
-        self.name.shrink_to_fit();
-        self.args.shrink_to_fit();
-        for arg in &mut self.args {
-            arg.shrink_to_fit();
-        }
-    }
-}
-
-impl fmt::Display for TypePathSegment {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.name)?;
-        if !self.args.is_empty() {
-            write!(f, "<")?;
-            for (idx, arg) in self.args.iter().enumerate() {
-                if idx > 0 {
-                    write!(f, ", ")?;
-                }
-                write!(f, "{arg}")?;
-            }
-            write!(f, ">")?;
-        }
-        Ok(())
-    }
-}
-
-#[derive(
-    Debug, Clone, PartialEq, Eq, wincode::SchemaRead, wincode::SchemaWrite, rg_memsize::MemorySize,
-)]
-pub enum GenericArg {
-    Type(#[wincode(with = "rg_wincode_utils::WincodeDynamic<TypeRef>")] TypeRef),
-    Lifetime(String),
-    Const(String),
-    AssocType {
-        name: Name,
-        #[wincode(with = "rg_wincode_utils::WincodeDynamic<Option<TypeRef>>")]
-        ty: Option<TypeRef>,
-    },
-    Unsupported(String),
-}
-
-impl GenericArg {
-    pub fn from_ast(
-        arg: ast::GenericArg,
-        line_index: &LineIndex,
-        interner: &mut NameInterner,
-    ) -> Self {
-        match arg {
+    fn from_ast(arg: &Self::AstNode, (line_index, interner): Self::Context<'_>) -> Self {
+        match arg.clone() {
             ast::GenericArg::AssocTypeArg(arg) => Self::AssocType {
                 name: arg
                     .name_ref()
@@ -541,7 +188,7 @@ impl GenericArg {
                     .unwrap_or_else(|| interner.intern("<missing>")),
                 ty: arg
                     .ty()
-                    .map(|ty| TypeRef::from_ast(ty, line_index, interner)),
+                    .map(|ty| TypeRef::from_ast(&ty, (line_index, &mut *interner))),
             },
             ast::GenericArg::ConstArg(arg) => Self::Const(normalized_syntax(&arg)),
             ast::GenericArg::LifetimeArg(arg) => arg
@@ -550,143 +197,37 @@ impl GenericArg {
                 .unwrap_or_else(|| Self::Unsupported(normalized_syntax(&arg))),
             ast::GenericArg::TypeArg(arg) => arg
                 .ty()
-                .map(|ty| TypeRef::from_ast(ty, line_index, interner))
+                .map(|ty| TypeRef::from_ast(&ty, (line_index, &mut *interner)))
                 .map(Self::Type)
                 .unwrap_or_else(|| Self::Unsupported(normalized_syntax(&arg))),
         }
     }
-
-    /// Extracts the syntax type from this argument when it is a type argument.
-    pub fn type_ref(&self) -> Option<&TypeRef> {
-        match self {
-            Self::Type(ty) => Some(ty),
-            Self::Lifetime(_) | Self::Const(_) | Self::AssocType { .. } | Self::Unsupported(_) => {
-                None
-            }
-        }
-    }
-
-    /// Returns true when this generic argument mentions one of the provided type parameter names.
-    pub fn mentions_type_param(&self, params: &[&str]) -> bool {
-        match self {
-            Self::Type(ty) => ty.mentions_type_param(params),
-            Self::AssocType { ty, .. } => {
-                ty.as_ref().is_some_and(|ty| ty.mentions_type_param(params))
-            }
-            Self::Lifetime(_) | Self::Const(_) | Self::Unsupported(_) => false,
-        }
-    }
-
-    pub fn shrink_to_fit(&mut self) {
-        match self {
-            Self::Type(ty) => ty.shrink_to_fit(),
-            Self::Lifetime(lifetime) | Self::Const(lifetime) | Self::Unsupported(lifetime) => {
-                lifetime.shrink_to_fit();
-            }
-            Self::AssocType { name, ty } => {
-                name.shrink_to_fit();
-                if let Some(ty) = ty {
-                    ty.shrink_to_fit();
-                }
-            }
-        }
-    }
 }
 
-impl fmt::Display for GenericArg {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Type(ty) => write!(f, "{ty}"),
-            Self::Lifetime(lifetime) => write!(f, "{lifetime}"),
-            Self::Const(value) => write!(f, "{value}"),
-            Self::AssocType { name, ty } => match ty {
-                Some(ty) => write!(f, "{name} = {ty}"),
-                None => write!(f, "{name}"),
-            },
-            Self::Unsupported(text) => write!(f, "<unsupported:{text}>"),
-        }
-    }
+pub(crate) fn type_bound_list_from_ast(
+    bound_list: Option<ast::TypeBoundList>,
+    line_index: &LineIndex,
+    interner: &mut NameInterner,
+) -> Vec<TypeBound> {
+    bound_list
+        .into_iter()
+        .flat_map(|bound_list| bound_list.bounds())
+        .map(|bound| type_bound_from_ast(bound, line_index, interner))
+        .collect()
 }
 
-#[derive(
-    Debug, Clone, PartialEq, Eq, wincode::SchemaRead, wincode::SchemaWrite, rg_memsize::MemorySize,
-)]
-pub enum TypeBound {
-    Trait(#[wincode(with = "rg_wincode_utils::WincodeDynamic<TypeRef>")] TypeRef),
-    Lifetime(String),
-    Unsupported(String),
-}
-
-impl TypeBound {
-    pub fn list_from_ast(
-        bound_list: Option<ast::TypeBoundList>,
-        line_index: &LineIndex,
-        interner: &mut NameInterner,
-    ) -> Vec<Self> {
-        bound_list
-            .into_iter()
-            .flat_map(|bound_list| bound_list.bounds())
-            .map(|bound| Self::from_ast(bound, line_index, interner))
-            .collect()
+fn type_bound_from_ast(
+    bound: ast::TypeBound,
+    line_index: &LineIndex,
+    interner: &mut NameInterner,
+) -> TypeBound {
+    if let Some(lifetime) = bound.lifetime() {
+        return TypeBound::Lifetime(normalized_syntax(&lifetime));
     }
 
-    fn from_ast(
-        bound: ast::TypeBound,
-        line_index: &LineIndex,
-        interner: &mut NameInterner,
-    ) -> Self {
-        if let Some(lifetime) = bound.lifetime() {
-            return Self::Lifetime(normalized_syntax(&lifetime));
-        }
-
-        if let Some(ty) = bound.ty() {
-            return Self::Trait(TypeRef::from_ast(ty, line_index, interner));
-        }
-
-        Self::Unsupported(normalized_syntax(&bound))
+    if let Some(ty) = bound.ty() {
+        return TypeBound::Trait(TypeRef::from_ast(&ty, (line_index, interner)));
     }
 
-    /// Returns true when this bound contains explicit generic arguments anywhere inside it.
-    pub fn has_generic_args(&self) -> bool {
-        match self {
-            Self::Trait(ty) => ty.has_generic_args(),
-            Self::Lifetime(_) | Self::Unsupported(_) => false,
-        }
-    }
-
-    /// Returns true when this bound mentions one of the provided type parameter names.
-    pub fn mentions_type_param(&self, params: &[&str]) -> bool {
-        match self {
-            Self::Trait(ty) => ty.mentions_type_param(params),
-            Self::Lifetime(_) | Self::Unsupported(_) => false,
-        }
-    }
-
-    pub fn shrink_to_fit(&mut self) {
-        match self {
-            Self::Trait(ty) => ty.shrink_to_fit(),
-            Self::Lifetime(lifetime) | Self::Unsupported(lifetime) => lifetime.shrink_to_fit(),
-        }
-    }
-}
-
-impl fmt::Display for TypeBound {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Trait(ty) => write!(f, "{ty}"),
-            Self::Lifetime(lifetime) => write!(f, "{lifetime}"),
-            Self::Unsupported(text) => write!(f, "<unsupported:{text}>"),
-        }
-    }
-}
-
-fn write_bounds(f: &mut fmt::Formatter<'_>, prefix: &str, bounds: &[TypeBound]) -> fmt::Result {
-    write!(f, "{prefix}")?;
-    for (idx, bound) in bounds.iter().enumerate() {
-        if idx > 0 {
-            write!(f, " + ")?;
-        }
-        write!(f, "{bound}")?;
-    }
-    Ok(())
+    TypeBound::Unsupported(normalized_syntax(&bound))
 }
