+++ crates/engine/item-tree/src/lib.rs
@@ -14,10 +14,12 @@ pub use self::{
     item::{
         BuiltinMacroItem, CfgAttrMacroUse, CfgSelectArmItem, CfgSelectArmPayload, ConstItem,
         Documentation, EnumItem, EnumVariantItem, ExternCrateItem, FieldItem, FieldKey, FieldList,
-        FunctionItem, FunctionQualifiers, GenericArg, GenericParams, ImplItem, ImportAlias,
-        ItemKind, ItemNode, ItemTag, ItemTreeId, ItemTreeRef, MacroCallItem, MacroDefinitionAttrs,
-        MacroDefinitionItem, MacroUseAttr, MacroUseSelector, ModuleItem, ModuleSource, Mutability,
-        ParamItem, ParamKind, StaticItem, StructItem, TraitItem, TypeAliasItem, TypeBound,
+        FromAst, FunctionItem, FunctionQualifiers, GenericArg, GenericParams, ImplItem,
+        ImplItemContext, ImportAlias, InnerDocs, ItemKind, ItemNode, ItemTag, ItemTreeId,
+        ItemTreeRef, MacroCallContext, MacroCallItem, MacroDefAst, MacroDefContext,
+        MacroDefinitionAttrs, MacroDefinitionItem, MacroRulesAst, MacroRulesContext, MacroUseAttr,
+        MacroUseSelector, MaybeFromAst, ModuleItem, ModuleSource, Mutability, OuterDocs, ParamItem,
+        ParamKind, StaticItem, StructItem, TraitItem, TraitItemContext, TypeAliasItem, TypeBound,
         TypePath, TypePathSegment, TypeRef, UnionItem, UseImport, UseImportKind, UseItem, UsePath,
         UsePathSegment, UsePathSegmentKind, VisibilityLevel, WherePredicate,
     },
